# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal value investing portfolio manager API",
        "title": "DeepValue API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/ingest/benchmark": {
            "post": {
                "description": "Fetches ETF benchmark prices (SPY, QQQ, etc.) from Tiingo - incremental updates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingestion"
                ],
                "summary": "Ingest benchmark prices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated tickers (defaults to SPY)",
                        "name": "ticker",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    }
                }
            }
        },
        "/admin/ingest/fundamentals": {
            "post": {
                "description": "Fetches fundamental data from SHARADAR/SF1 and upserts into financial_metrics table",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingestion"
                ],
                "summary": "Ingest financial metrics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated tickers (defaults to all companies in DB)",
                        "name": "ticker",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "ARQ,MRQ",
                        "description": "Comma-separated dimensions",
                        "name": "dimension",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Fetch all history (default: incremental)",
                        "name": "full",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    }
                }
            }
        },
        "/admin/ingest/prices": {
            "post": {
                "description": "Fetches daily prices for stocks from Tiingo (rate limited: 50/hour) - incremental updates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingestion"
                ],
                "summary": "Ingest daily stock prices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated tickers (defaults to batch of stocks needing prices)",
                        "name": "ticker",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max tickers to fetch (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Consider data stale after N days (default 3)",
                        "name": "stale_days",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Days to wait before retrying tickers with no data (default 7)",
                        "name": "retry_days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    }
                }
            }
        },
        "/admin/ingest/sp500": {
            "post": {
                "description": "Fetches S\u0026P 500 membership history from SHARADAR/SP500 for point-in-time backtesting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingestion"
                ],
                "summary": "Ingest S\u0026P 500 membership history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/ingest/tickers": {
            "post": {
                "description": "Fetches company metadata from SHARADAR/TICKERS and upserts into the companies table",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingestion"
                ],
                "summary": "Ingest company tickers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated tickers (defaults to all)",
                        "name": "ticker",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.IngestResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.IngestResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "elapsed": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}